requirements.txt

# Testing
pytest>=8.0.0
pytest-qt>=4.4.0
pytest-asyncio>=0.25.0
pytest-cov>=6.0.0
pytest-mock>=3.12.0

# GUI
PyQt6>=6.6.1
pyqtgraph>=0.13.3

# Hardware Communication
pyserial>=3.5
minimalmodbus>=2.1.1

# Data Processing
numpy>=1.26.0
pandas>=2.1.0

# Utilities
logging>=0.4.9.6
psutil>=5.9.0  # For performance monitoring
typing-extensions>=4.9.0

# Development
black>=24.1.0  # Code formatting
flake8>=7.0.0  # Linting
mypy>=1.8.0    # Type checking